.claude/skills/twitter-post/requirements.txt
# Twitter Post Skill Dependencies

tweepy>=4.14.0
python-dotenv>=1.0.0
